// types.d.ts
export interface post{
    _id:string,
    author:{
        name:string,
        image:string
    },
    description:string,
    _createdAt:string,
    slug:{
        current:string
    },
    title:string,
    mainImage:{
        asset:{
          _ref:string
        }
    },
    body:[object],
    comments:comment[]
}
export interface comment{
    approved:boolean,
    comment:string,
    email:string,
    name:string,
    post:{
        _ref:string,
        _type:string
    },
    _createdAt:string,
    _id:string,
    _rev:string,
    _type:string,
    _updatedAt:string
}